DRY code, clean up
