+++ server/queryHandlers/ScanInstances.js
@@ -4,14 +4,13 @@
 let resourceProvider = require('modules/resourceProvider');
 let co = require('co');
 
-module.exports = function ScanInstancesQueryHandler(query) {
+module.exports = function ScanInstancesQueryHandler({ accountName, filter }) {
   return co(function* () {
     // Create an instance of the resource to work with based on the resource
     // descriptor and AWS account name.
-    let parameters = { accountName: query.accountName };
-    let resource = yield resourceProvider.getInstanceByName('instances', parameters);
+    let resource = yield resourceProvider.getInstanceByName('instances', { accountName });
 
     // Scan resource items
-    return resource.all({ filter: query.filter });
+    return resource.all({ filter });
   });
 };
